requirements.txt
fastapi==0.109.1
uvicorn==0.24.0
tensorflow==2.16.1
numpy==1.26.4
pillow==10.4.0
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
python-multipart==0.0.19
locust==2.17.0
psutil==5.9.6
boto3==1.29.7

GitPython~=3.1.45
gitdb~=4.0.12
typing_extensions~=4.12.2
h11~=0.16.0
cffi~=1.17.1
defusedxml~=0.7.1
ipython~=9.6.0
pip~=25.3
attrs~=25.3.0
protobuf~=4.25.8
cryptography~=44.0.3
Booktype~=1.5
pyzmq~=27.1.0
colorama~=0.4.6
gevent~=25.9.1
pytest~=8.4.1
tornado~=6.5.2
absl-py~=2.3.1
setuptools~=80.9.0
comm~=0.2.3
gast~=0.6.0
grpcio~=1.75.1
h5py~=3.15.1
parso~=0.8.5
jedi~=0.19.2
pytz~=2025.2
rich~=13.9.4
Pygments~=2.19.2
markdown-it-py~=4.0.0
toml~=0.10.2
sniffio~=1.3.1
botocore~=1.32.7
s3transfer~=0.8.0
jmespath~=1.0.1
libclang~=18.1.1
MarkupSafe~=3.0.2
Werkzeug~=3.1.3
click~=8.1.7
Jinja2~=3.1.4
blinker~=1.9.0
itsdangerous~=2.2.0
smmap~=5.0.2
pywin32~=311
keras~=3.11.3
ml-dtypes~=0.3.2
optree~=0.17.0
packaging~=24.2
scipy~=1.16.2
scikit-learn~=1.5.0
requests~=2.32.4
namex~=0.1.0
mdurl~=0.1.2
charset-normalizer~=3.4.2
numba~=0.62.1
six~=1.16.0
google-pasta~=0.2.0
wheel~=0.45.1
altair~=5.5.0
jsonschema~=4.25.1
narwhals~=2.12.0
pyarrow~=22.0.0
referencing~=0.37.0
traitlets~=5.14.3
schema~=0.7.8
greenlet~=3.2.4
dnspython~=2.8.0
coverage~=7.6.12
msgpack~=1.1.2
Flask~=3.1.0
urllib3~=2.0.7
geventhttpclient~=2.3.5
ConfigArgParse~=1.7.1
Flask-BasicAuth~=0.2.0
flask-cors~=5.0.1
roundrobin~=0.0.4
SQLAlchemy~=2.0.36
lxml~=5.3.1
python-dateutil~=2.9.0.post0
tenacity~=8.5.0
scikit-image~=0.25.2
pydeck~=0.9.1
certifi~=2025.1.31
debugpy~=1.8.17
Django~=5.2.5
decorator~=5.2.1
starlette~=0.35.1
email-validator~=2.3.0
pydantic~=2.10.3
pydantic_core~=2.27.1
websockets~=15.0.1
anyio~=4.9.0
ipython_pygments_lexers~=1.1.1
stack-data~=0.6.3
jupyter_client~=8.6.3
prompt_toolkit~=3.0.52
wcwidth~=0.2.14
tzdata~=2025.2
cycler~=0.12.1
httptools~=0.7.1
watchfiles~=1.1.1
Markdown~=3.9
PyYAML~=6.0.3
annotated-types~=0.7.0
idna~=3.10
simplejson~=3.20.1
watchdog~=4.0.2
asttokens~=3.0.0
contourpy~=1.3.2
executing~=2.2.1
fonttools~=4.58.0
path~=17.1.0
ipykernel~=7.0.1
matplotlib-inline~=0.1.7
jupyter_core~=5.9.1
pure_eval~=0.2.3
pycparser~=2.22
pyparsing~=3.2.3
httpx~=0.28.1
streamlit~=1.37.0
cachetools~=5.5.2
graphviz~=0.20.1
termcolor~=3.1.0
webcolors~=24.11.1
jsonpointer~=3.0.0
uri-template~=1.3.0
isoduration~=20.11.0
fqdn~=1.5.1
rfc3339-validator~=0.1.4
jsonschema-specifications~=2025.9.1
kiwisolver~=1.4.8
opt_einsum~=3.4.0
flatbuffers~=25.9.23
wrapt~=1.17.3
astunparse~=1.6.3
tensorboard~=2.16.2
platformdirs~=4.5.0
brotli~=1.2.0